requirements.txt
matplotlib==3.6.1
numpy==1.23.0
open_metric_learning==0.4.5
pandas==1.5.2
Pillow==9.5.0
Requests==2.31.0
scikit_learn==1.3.0
tensorflow==2.9.0
torch==1.13.0
faiss==1.7.4